Move asset & pack add to Dialog bottom actions
