+++ newIDE/app/src/AssetStore/NewObjectDialog.js
@@ -7,6 +7,7 @@ import ListIcon from '../UI/ListIcon';
 import Subheader from '../UI/Subheader';
 import { List, ListItem } from '../UI/List';
 import {
+  enumerateObjects,
   enumerateObjectTypes,
   type EnumeratedObjectMetadata,
 } from '../ObjectsList/EnumerateObjects';
@@ -19,10 +20,20 @@ import {
   type ChooseResourceFunction,
 } from '../ResourcesList/ResourceSource';
 import { type ResourceExternalEditor } from '../ResourcesList/ResourceExternalEditor.flow';
-import { sendNewObjectCreated } from '../Utils/Analytics/EventSender';
+import {
+  sendAssetAddedToProject,
+  sendNewObjectCreated,
+} from '../Utils/Analytics/EventSender';
 import PreferencesContext from '../MainFrame/Preferences/PreferencesContext';
 import ScrollView from '../UI/ScrollView';
 import useDismissableTutorialMessage from '../Hints/useDismissableTutorialMessage';
+import RaisedButton from '../UI/RaisedButton';
+import { AssetStoreContext } from './AssetStoreContext';
+import { AssetPackDialog } from './AssetPackDialog';
+import { installAsset } from './InstallAsset';
+import EventsFunctionsExtensionsContext from '../EventsFunctionsExtensionsLoader/EventsFunctionsExtensionsContext';
+import { useResourceFetcher } from '../ProjectsStorage/ResourceFetcher';
+import { showErrorBox } from '../UI/Messages/MessageBox';
 
 const ObjectListItem = ({
   objectMetadata,
@@ -115,77 +126,195 @@ export default function NewObjectDialog({
     'intro-object-types'
   );
 
+  const { searchResults, navigationState } = React.useContext(
+    AssetStoreContext
+  );
+  const {
+    openedAssetPack,
+    openedAssetShortHeader,
+  } = navigationState.getCurrentPage();
+  const [
+    isAssetPackDialogInstallOpen,
+    setIsAssetPackDialogInstallOpen,
+  ] = React.useState(false);
+  const { containerObjectsList } = enumerateObjects(project, objectsContainer);
+  const addedAssetIds = containerObjectsList
+    .map(({ object }) => object.getAssetStoreId())
+    .filter(Boolean);
+  const [
+    isAssetBeingInstalled,
+    setIsAssetBeingInstalled,
+  ] = React.useState<boolean>(false);
+  const resourcesFetcher = useResourceFetcher();
+  const eventsFunctionsExtensionsState = React.useContext(
+    EventsFunctionsExtensionsContext
+  );
+  const isAssetAddedToScene =
+    openedAssetShortHeader && addedAssetIds.includes(openedAssetShortHeader.id);
+
+  const onInstallAsset = React.useCallback(
+    () => {
+      setIsAssetBeingInstalled(true);
+      (async () => {
+        if (!openedAssetShortHeader) return;
+        try {
+          const installOutput = await installAsset({
+            assetShortHeader: openedAssetShortHeader,
+            eventsFunctionsExtensionsState,
+            project,
+            objectsContainer,
+            events,
+          });
+          sendAssetAddedToProject({
+            id: openedAssetShortHeader.id,
+            name: openedAssetShortHeader.name,
+          });
+          console.log('Asset successfully installed.');
+
+          installOutput.createdObjects.forEach(object => {
+            onObjectAddedFromAsset(object);
+          });
+
+          await resourcesFetcher.ensureResourcesAreFetched(project);
+        } catch (error) {
+          console.error('Error while installing the asset:', error);
+          showErrorBox({
+            message: `There was an error while installing the asset "${
+              openedAssetShortHeader.name
+            }". Verify your internet connection or try again later.`,
+            rawError: error,
+            errorId: 'install-asset-error',
+          });
+        }
+
+        setIsAssetBeingInstalled(false);
+      })();
+    },
+    [
+      resourcesFetcher,
+      eventsFunctionsExtensionsState,
+      project,
+      objectsContainer,
+      events,
+      onObjectAddedFromAsset,
+      openedAssetShortHeader,
+    ]
+  );
+
+  const mainAction = openedAssetPack ? (
+    <RaisedButton
+      key="add-all-assets"
+      primary
+      label={<Trans>Add all assets to my scene</Trans>}
+      onClick={() => setIsAssetPackDialogInstallOpen(true)}
+      disabled={!searchResults || searchResults.length === 0}
+    />
+  ) : openedAssetShortHeader ? (
+    <RaisedButton
+      key="add-asset"
+      primary={!isAssetAddedToScene}
+      label={
+        isAssetBeingInstalled ? (
+          <Trans>Adding...</Trans>
+        ) : isAssetAddedToScene ? (
+          <Trans>Add again</Trans>
+        ) : (
+          <Trans>Add to the scene</Trans>
+        )
+      }
+      onClick={onInstallAsset}
+      disabled={isAssetBeingInstalled}
+      id="add-asset-button"
+    />
+  ) : (
+    undefined
+  );
+
   return (
-    <Dialog
-      title={<Trans>Add a new object</Trans>}
-      secondaryActions={[<HelpButton helpPagePath="/objects" key="help" />]}
-      actions={[
-        <FlatButton
-          key="close"
-          label={<Trans>Close</Trans>}
-          primary={false}
-          onClick={onClose}
-          id="close-button"
-        />,
-      ]}
-      onRequestClose={onClose}
-      open
-      flexBody
-      noMargin
-      fullHeight
-      id="new-object-dialog"
-    >
-      <Column noMargin expand>
-        <Tabs value={currentTab} onChange={setCurrentTab}>
-          <Tab label={<Trans>Asset Store</Trans>} value="asset-store" />
-          <Tab
-            label={<Trans>New object from scratch</Trans>}
-            value="new-object"
-          />
-        </Tabs>
-        {currentTab === 'asset-store' && (
-          <AssetStore
-            focusOnMount
-            project={project}
-            objectsContainer={objectsContainer}
-            events={events}
-            layout={layout}
-            onChooseResource={onChooseResource}
-            resourceSources={resourceSources}
-            resourceExternalEditors={resourceExternalEditors}
-            onObjectAddedFromAsset={onObjectAddedFromAsset}
-          />
-        )}
-        {currentTab === 'new-object' && (
-          <ScrollView>
-            {DismissableTutorialMessage && (
-              <Line>
-                <Column expand>{DismissableTutorialMessage}</Column>
-              </Line>
-            )}
-            <List>
-              {Object.keys(objectsByCategory).map(category => {
-                const categoryObjectMetadata = objectsByCategory[category];
-                return (
-                  <React.Fragment key={category}>
-                    <Subheader>{category}</Subheader>
-                    {categoryObjectMetadata.map(objectMetadata => (
-                      <ObjectListItem
-                        key={objectMetadata.name}
-                        objectMetadata={objectMetadata}
-                        onClick={() => {
-                          sendNewObjectCreated(objectMetadata.name);
-                          onCreateNewObject(objectMetadata.name);
-                        }}
-                      />
-                    ))}
-                  </React.Fragment>
-                );
-              })}
-            </List>
-          </ScrollView>
-        )}
-      </Column>
-    </Dialog>
+    <>
+      <Dialog
+        title={<Trans>Add a new object</Trans>}
+        secondaryActions={[<HelpButton helpPagePath="/objects" key="help" />]}
+        actions={[
+          <FlatButton
+            key="close"
+            label={<Trans>Close</Trans>}
+            primary={false}
+            onClick={onClose}
+            id="close-button"
+          />,
+          mainAction,
+        ]}
+        onRequestClose={onClose}
+        onApply={
+          openedAssetPack
+            ? () => setIsAssetPackDialogInstallOpen(true)
+            : openedAssetShortHeader
+            ? onInstallAsset
+            : undefined
+        }
+        open
+        flexBody
+        noMargin
+        fullHeight
+        id="new-object-dialog"
+      >
+        <Column noMargin expand>
+          <Tabs value={currentTab} onChange={setCurrentTab}>
+            <Tab label={<Trans>Asset Store</Trans>} value="asset-store" />
+            <Tab
+              label={<Trans>New object from scratch</Trans>}
+              value="new-object"
+            />
+          </Tabs>
+          {currentTab === 'asset-store' && <AssetStore project={project} />}
+          {currentTab === 'new-object' && (
+            <ScrollView>
+              {DismissableTutorialMessage && (
+                <Line>
+                  <Column expand>{DismissableTutorialMessage}</Column>
+                </Line>
+              )}
+              <List>
+                {Object.keys(objectsByCategory).map(category => {
+                  const categoryObjectMetadata = objectsByCategory[category];
+                  return (
+                    <React.Fragment key={category}>
+                      <Subheader>{category}</Subheader>
+                      {categoryObjectMetadata.map(objectMetadata => (
+                        <ObjectListItem
+                          key={objectMetadata.name}
+                          objectMetadata={objectMetadata}
+                          onClick={() => {
+                            sendNewObjectCreated(objectMetadata.name);
+                            onCreateNewObject(objectMetadata.name);
+                          }}
+                        />
+                      ))}
+                    </React.Fragment>
+                  );
+                })}
+              </List>
+            </ScrollView>
+          )}
+        </Column>
+      </Dialog>
+      {isAssetPackDialogInstallOpen && searchResults && openedAssetPack && (
+        <AssetPackDialog
+          assetPack={openedAssetPack}
+          assetShortHeaders={searchResults}
+          addedAssetIds={addedAssetIds}
+          onClose={() => setIsAssetPackDialogInstallOpen(false)}
+          onAssetsAdded={() => {
+            setIsAssetPackDialogInstallOpen(false);
+          }}
+          project={project}
+          objectsContainer={objectsContainer}
+          events={events}
+          onObjectAddedFromAsset={onObjectAddedFromAsset}
+        />
+      )}
+      {resourcesFetcher.renderResourceFetcherDialog()}
+    </>
   );
 }
